rename config files
