+++ app/config/index.js
@@ -1,5 +1,5 @@
-const development = require('./development');
-const production = require('./production');
+const development = require('./development.config');
+const production = require('./production.config');
 
 module.exports = process.env
     .NODE_ENV == 'production' ? production : development;
