+++ server.js
@@ -1,28 +1,14 @@
-const bodyParser = require('body-parser');
-const express = require('express');
-const methodOverride = require('method-override');
 const mongoose = require('mongoose');
-const morgan = require('morgan');
-
+const app = require('./app');
 const config = require('./app/config');
-const router = require('./app/config/router');
-
-const app = express();
-
-// log to console in dev
-app.use(morgan('dev'));
-// parse application/x-www-form-urlencoded
-app.use(bodyParser.urlencoded({'extended': true}));
-// parse application/json
-app.use(bodyParser.json());
-app.use(router);
-app.use(methodOverride);
 
 // connecting MongoDB using mongoose to our application
 mongoose.set('useCreateIndex', true);
 mongoose.connect(
   config.db,
-  {useNewUrlParser: true});
+  {useNewUrlParser: true},
+);
+
 // this callback will be triggered on successful connection
 mongoose.connection.on('connected', () => {
   console.log(`Mongoose connected successfully to ${config.db}`);
